[DIET-55-auth-controller] Id Integer -> Long in every model

+++ dyeti-backend/src/main/java/ki/agh/dyeti/model/PlansProducts.java
@@ -19,9 +19,9 @@ import lombok.NoArgsConstructor;
 @IdClass(PlansProductsId.class)
 public class PlansProducts {
 
-  @Id private Integer planId;
+  @Id private Long planId;
 
-  @Id private Integer productId;
+  @Id private Long productId;
 
   private Double quantity;
 }
