+++ dyeti-backend/src/main/java/ki/agh/dyeti/model/RecipesProducts.java
@@ -19,9 +19,9 @@ import lombok.NoArgsConstructor;
 @IdClass(RecipesProductsId.class)
 public class RecipesProducts {
 
-  @Id private Integer recipeId;
+  @Id private Long recipeId;
 
-  @Id private Integer productId;
+  @Id private Long productId;
 
   private Double quantity;
   private PortionType portionType;
